+++ Assets/OxGKit/Utilities/Scripts/Runtime/Timer/RealTimer.cs
@@ -18,7 +18,7 @@ namespace OxGKit.Utilities.Timer
         private float _lastTickTime;
 
         private float _mark;
-        private float _speed;
+        private float _timeSpeed;
 
         public RealTimer()
         {
@@ -54,7 +54,7 @@ namespace OxGKit.Utilities.Timer
             this._tickTime = 0.0f;
             this._lastTickTime = 0.0f;
             this._mark = 0.0f;
-            this._speed = 1.0f;
+            this._timeSpeed = 1.0f;
         }
 
         public float GetRealTime()
@@ -66,7 +66,7 @@ namespace OxGKit.Utilities.Timer
         public float GetTime()
         {
             if (!this._playing) return this._pauseTime - this._intervalTime;
-            return (this.GetRealTime() - this._intervalTime) * this._speed;
+            return (this.GetRealTime() - this._intervalTime) * this._timeSpeed;
         }
 
         public void Pause()
@@ -90,6 +90,7 @@ namespace OxGKit.Utilities.Timer
 
         public void Play()
         {
+            if (this._playing) return;
             this._intervalTime += this.GetRealTime() - this._pauseTime;
             this._playing = true;
         }
@@ -108,10 +109,10 @@ namespace OxGKit.Utilities.Timer
         /// <summary>
         /// 設置要計時的秒數
         /// </summary>
-        /// <param name="time"></param>
-        public void SetTimer(float time)
+        /// <param name="timeSeconds"></param>
+        public void SetTimer(float timeSeconds)
         {
-            this._timerTime = time;
+            this._timerTime = timeSeconds;
             this._triggerTime = this.GetTime() + this._timerTime;
         }
 
@@ -132,8 +133,8 @@ namespace OxGKit.Utilities.Timer
         /// <returns></returns>
         public bool IsTimerTimeout()
         {
-            if (this.GetTime() < this._triggerTime) return false;
-            return true;
+            if (this.GetTime() > this._triggerTime) return true;
+            return false;
         }
 
         /// <summary>
@@ -153,10 +154,10 @@ namespace OxGKit.Utilities.Timer
         /// <summary>
         /// 設置 Tick 時間, 當 TickTimeout 時還會持續循環 Tick
         /// </summary>
-        /// <param name="tick"></param>
-        public void SetTick(float tick)
+        /// <param name="tickSeconds"></param>
+        public void SetTick(float tickSeconds)
         {
-            this._tickTime = tick;
+            this._tickTime = tickSeconds;
             this._lastTickTime = this.GetTime() + this._tickTime;
         }
 
@@ -187,9 +188,12 @@ namespace OxGKit.Utilities.Timer
         public bool IsTickTimeout()
         {
             float time = this.GetTime();
-            if (time < this._lastTickTime) return false;
-            this._lastTickTime = time + this._tickTime;
-            return true;
+            if (time > this._lastTickTime)
+            {
+                this._lastTickTime = time + this._tickTime;
+                return true;
+            }
+            return false;
         }
 
         /// <summary>
@@ -231,22 +235,22 @@ namespace OxGKit.Utilities.Timer
         {
             float time = this.GetTime();
             if (time == this._mark || time < this._mark) return 0.0f;
-            return (time - this._mark);
+            return time - this._mark;
         }
         #endregion
 
         /// <summary>
         /// 設置時間運轉速度
         /// </summary>
-        /// <param name="speed"></param>
-        public void SetSpeed(float speed)
+        /// <param name="timeSpeed"></param>
+        public void SetTimeSpeed(float timeSpeed)
         {
-            this._speed = speed;
+            this._timeSpeed = timeSpeed;
         }
 
-        public float GetSpeed()
+        public float GetTimeSpeed()
         {
-            return this._speed;
+            return this._timeSpeed;
         }
     }
 }